refactor(homepage): 내부 desktop/mobile 폴더 구조 변경 (#280)
