+++ apps/homepage/src/app/(with-header)/(with-footer)/about-us/_mobile/_components/AboutUsMobileManagementAccordionItem.tsx
